refactor: rename markup to template
