+++ src/components/template/index.js
